refactor(blorktools): state and reversal helper
