+++ packages/blorktools/src/asset_debugger/util/css3d/css3d-bounce-helper.js
@@ -1,6 +1,18 @@
 import { classifyAnimation, classifyTransition } from "../animation-classifier";
-import { animationProperties, animationStack, batchTimeWindow, currentAnimationBatch, isCapturingAnimations, lastAnimationTime, lastBatchTime, pushAnimationBatch, resetAnimationState, resetCurrentAniamtionBatch, setLastAnimationTime, setLastBatchTime } from "./css3d-util";
-
+import { 
+    animationProperties,
+    animationStack,
+    batchTimeWindow,
+    currentAnimationBatch,
+    isCapturingAnimations,
+    lastAnimationTime,
+    lastBatchTime,
+    pushAnimationBatch,
+    resetAnimationState,
+    resetCurrentAniamtionBatch,
+    setLastAnimationTime,
+    setLastBatchTime
+} from "./css3d-state";
 
 /**
  * Set up bounce animation tracking for the iframe
